gapd-2021/lecture-8/transaction-stream/requirements.txt
gqlalchemy==1.0.5
kafka-python==2.0.2
